requirements.txt
livekit-agents>=0.12.10
livekit-plugins-deepgram>=0.6.16
livekit-plugins-elevenlabs>=0.7.1
livekit-plugins-google>=0.10.1
livekit-plugins-openai[vertex]>=0.10.10
livekit-plugins-silero>=0.7.4
livekit-plugins-rag>=0.2.3
python-dotenv~=1.0
aiofile~=3.8.8